feat(broker): 1. Service refactor for broker
