+++ landuse_app/__init__.py
@@ -1,9 +1,1 @@
-"""Main module."""
 
-from .config import config
-from .fastapi_init import app
-
-__all__ = [
-    "app",
-    "config",
-]
