+++ landuse_app/logic/__init__.py
@@ -1,8 +1,1 @@
-"""Logic module."""
 
-from .landuse_service import LanduseService, landuse_service
-
-__all__ = [
-    "LanduseService",
-    "landuse_service",
-]
