+++ landuse_app/logic/api/__init__.py
@@ -1,3 +1,1 @@
-from .urban_db_api_client import urban_db_api
 
-__all__ = ["urban_db_api"]
