+++ landuse_app/logic/helpers/__init__.py
@@ -1,23 +1,1 @@
-"""Helper functions module."""
 
-from .indicators_service import IndicatorsService
-from .renovation_potential import (
-    get_projects_context_renovation_potential,
-    get_projects_context_urbanization_level,
-    get_projects_renovation_potential,
-    get_projects_urbanization_level,
-)
-from .urban_api_access import (
-    get_projects_base_scenario_id,
-    get_projects_territory,
-)
-
-__all__ = [
-    "get_projects_renovation_potential",
-    "get_projects_context_renovation_potential",
-    "get_projects_urbanization_level",
-    "get_projects_context_urbanization_level",
-    "get_projects_territory",
-    "get_projects_base_scenario_id",
-    "IndicatorsService",
-]
